# Make delayed sampling optional