--- a/sampler.go
+++ b/sampler.go
@@ -42,7 +42,7 @@ type Sampler interface {
 
 	// Equal checks if the `other` sampler is functionally equivalent
 	// to this sampler.
-	// TODO (breaking change) remove this function. See AdaptiveSampler.Equals for explanation.
+	// TODO (breaking change) remove this function. See PerOperationSampler.Equals for explanation.
 	Equal(other Sampler) bool
 }
 
@@ -305,7 +305,7 @@ func (s *GuaranteedThroughputProbabilisticSampler) Close() {
 
 // Equal implements Equal() of Sampler.
 func (s *GuaranteedThroughputProbabilisticSampler) Equal(other Sampler) bool {
-	// NB The Equal() function is expensive and will be removed. See AdaptiveSampler.Equal() for
+	// NB The Equal() function is expensive and will be removed. See PerOperationSampler.Equal() for
 	// more information.
 	return false
 }
@@ -321,9 +321,9 @@ func (s *GuaranteedThroughputProbabilisticSampler) update(lowerBound, samplingRa
 
 // -----------------------
 
-// AdaptiveSampler is a delegating sampler that applies GuaranteedThroughputProbabilisticSampler
+// PerOperationSampler is a delegating sampler that applies GuaranteedThroughputProbabilisticSampler
 // on a per-operation basis.
-type AdaptiveSampler struct {
+type PerOperationSampler struct {
 	sync.RWMutex
 
 	samplers       map[string]*GuaranteedThroughputProbabilisticSampler
@@ -331,12 +331,12 @@ type AdaptiveSampler struct {
 	lowerBound     float64
 	maxOperations  int
 
-	// see description in AdaptiveSamplerParams
+	// see description in PerOperationSamplerParams
 	operationNameLateBinding bool
 }
 
-// AdaptiveSamplerParams defines parameters when creating an adaptive sampler.
-type AdaptiveSamplerParams struct {
+// PerOperationSamplerParams defines parameters when creating an adaptive sampler.
+type PerOperationSamplerParams struct {
 	// Max number of operations that will be tracked. Other operations will be given default strategy.
 	MaxOperations int
 
@@ -352,17 +352,17 @@ type AdaptiveSamplerParams struct {
 	Strategies *sampling.PerOperationSamplingStrategies
 }
 
-// NewAdaptiveSampler returns a new AdaptiveSampler.
-// TODO (breaking change) remove error from return value
-func NewAdaptiveSampler(strategies *sampling.PerOperationSamplingStrategies, maxOperations int) (*AdaptiveSampler, error) {
-	return NewAdaptiveSamplerWithParams(AdaptiveSamplerParams{
+// NewAdaptiveSampler returns a new PerOperationSampler.
+// Deprecated: please use NewPerOperationSampler.
+func NewAdaptiveSampler(strategies *sampling.PerOperationSamplingStrategies, maxOperations int) (*PerOperationSampler, error) {
+	return NewPerOperationSampler(PerOperationSamplerParams{
 		MaxOperations: maxOperations,
 		Strategies:    strategies,
 	}), nil
 }
 
-// NewAdaptiveSamplerWithParams returns a new AdaptiveSampler.
-func NewAdaptiveSamplerWithParams(params AdaptiveSamplerParams) *AdaptiveSampler {
+// NewPerOperationSampler returns a new PerOperationSampler.
+func NewPerOperationSampler(params PerOperationSamplerParams) *PerOperationSampler {
 	samplers := make(map[string]*GuaranteedThroughputProbabilisticSampler)
 	for _, strategy := range params.Strategies.PerOperationStrategies {
 		sampler := newGuaranteedThroughputProbabilisticSampler(
@@ -371,7 +371,7 @@ func NewAdaptiveSamplerWithParams(params AdaptiveSamplerParams) *AdaptiveSampler
 		)
 		samplers[strategy.Operation] = sampler
 	}
-	return &AdaptiveSampler{
+	return &PerOperationSampler{
 		samplers:                 samplers,
 		defaultSampler:           newProbabilisticSampler(params.Strategies.DefaultSamplingProbability),
 		lowerBound:               params.Strategies.DefaultLowerBoundTracesPerSecond,
@@ -382,11 +382,11 @@ func NewAdaptiveSamplerWithParams(params AdaptiveSamplerParams) *AdaptiveSampler
 
 // IsSampled is not used and only exists to match Sampler V1 API.
 // TODO (breaking change) remove when upgrading everything to SamplerV2
-func (s *AdaptiveSampler) IsSampled(id TraceID, operation string) (bool, []Tag) {
+func (s *PerOperationSampler) IsSampled(id TraceID, operation string) (bool, []Tag) {
 	return false, nil
 }
 
-func (s *AdaptiveSampler) trySampling(span *Span, operationName string) (bool, []Tag) {
+func (s *PerOperationSampler) trySampling(span *Span, operationName string) (bool, []Tag) {
 	samplerV1 := s.getSamplerForOperation(operationName)
 	var sampled bool
 	var tags []Tag
@@ -397,28 +397,28 @@ func (s *AdaptiveSampler) trySampling(span *Span, operationName string) (bool, [
 }
 
 // OnCreateSpan implements OnCreateSpan of SamplerV2.
-func (s *AdaptiveSampler) OnCreateSpan(span *Span) SamplingDecision {
+func (s *PerOperationSampler) OnCreateSpan(span *Span) SamplingDecision {
 	sampled, tags := s.trySampling(span, span.OperationName())
 	return SamplingDecision{Sample: sampled, Retryable: s.operationNameLateBinding, Tags: tags}
 }
 
 // OnSetOperationName implements OnSetOperationName of SamplerV2.
-func (s *AdaptiveSampler) OnSetOperationName(span *Span, operationName string) SamplingDecision {
+func (s *PerOperationSampler) OnSetOperationName(span *Span, operationName string) SamplingDecision {
 	sampled, tags := s.trySampling(span, operationName)
 	return SamplingDecision{Sample: sampled, Retryable: false, Tags: tags}
 }
 
 // OnSetTag implements OnSetTag of SamplerV2.
-func (s *AdaptiveSampler) OnSetTag(span *Span, key string, value interface{}) SamplingDecision {
+func (s *PerOperationSampler) OnSetTag(span *Span, key string, value interface{}) SamplingDecision {
 	return SamplingDecision{Sample: false, Retryable: true}
 }
 
 // OnFinishSpan implements OnFinishSpan of SamplerV2.
-func (s *AdaptiveSampler) OnFinishSpan(span *Span) SamplingDecision {
+func (s *PerOperationSampler) OnFinishSpan(span *Span) SamplingDecision {
 	return SamplingDecision{Sample: false, Retryable: true}
 }
 
-func (s *AdaptiveSampler) getSamplerForOperation(operation string) Sampler {
+func (s *PerOperationSampler) getSamplerForOperation(operation string) Sampler {
 	s.RLock()
 	sampler, ok := s.samplers[operation]
 	if ok {
@@ -444,7 +444,7 @@ func (s *AdaptiveSampler) getSamplerForOperation(operation string) Sampler {
 }
 
 // Close invokes Close on all underlying samplers.
-func (s *AdaptiveSampler) Close() {
+func (s *PerOperationSampler) Close() {
 	s.Lock()
 	defer s.Unlock()
 	for _, sampler := range s.samplers {
@@ -455,16 +455,16 @@ func (s *AdaptiveSampler) Close() {
 
 // Equal is not used.
 // TODO (breaking change) remove this in the future
-func (s *AdaptiveSampler) Equal(other Sampler) bool {
-	// NB The Equal() function is overly expensive for AdaptiveSampler since it's composed of multiple
+func (s *PerOperationSampler) Equal(other Sampler) bool {
+	// NB The Equal() function is overly expensive for PerOperationSampler since it's composed of multiple
 	// samplers which all need to be initialized before this function can be called for a comparison.
-	// Therefore, AdaptiveSampler uses the update() function to only alter the samplers that need
+	// Therefore, PerOperationSampler uses the update() function to only alter the samplers that need
 	// changing. Hence this function always returns false so that the update function can be called.
 	// Once the Equal() function is removed from the Sampler API, this will no longer be needed.
 	return false
 }
 
-func (s *AdaptiveSampler) update(strategies *sampling.PerOperationSamplingStrategies) {
+func (s *PerOperationSampler) update(strategies *sampling.PerOperationSamplingStrategies) {
 	s.Lock()
 	defer s.Unlock()
 	newSamplers := map[string]*GuaranteedThroughputProbabilisticSampler{}