--- a/sampler_remote_test.go
+++ b/sampler_remote_test.go
@@ -303,9 +303,10 @@ func TestRemotelyControlledSampler_updateSamplerFromAdaptiveSampler(t *testing.T
 		DefaultSamplingProbability:       testDefaultSamplingProbability,
 		DefaultLowerBoundTracesPerSecond: 1.0,
 	}
-
-	adaptiveSampler, err := NewAdaptiveSampler(strategies, testDefaultMaxOperations)
-	require.NoError(t, err)
+	adaptiveSampler := NewAdaptiveSamplerWithParams(AdaptiveSamplerParams{
+		MaxOperations: testDefaultMaxOperations,
+		Strategies:    strategies,
+	})
 
 	// Overwrite the sampler with an adaptive sampler
 	remoteSampler.sampler = adaptiveSampler