Part of #449.

Recent changes to `adaptiveSampler` introduced a potential performance degradation, addressed here via a new opt-in option `OperationNameLateBinding`:

> Opt-in feature for applications that require late binding of span name via explicit call to `SetOperationName`. When this feature is enabled, the sampler will return `retryable=true` from `OnCreateSpan()`, thus leaving the sampling decision as non-final (and the span as writeable). This may lead to degraded performance in applications that always provide the correct span name on trace creation.

The PR also renames `AdaptiveSampler` to `PerOperationSampler`, to better reflect its purpose. The public name AdaptiveSampler has not been released yet, the previous release had it private `adaptiveSampler`.

## Benchmarks

Legend:
  * `children=false` means that every span starts a new trace (and getting sampled)
  * `children=true` means that most spans are created as a child of the previous span
  * Standalone `NeverSample/AlwaysSample` means using Const sampler
  * Adaptive(NeverSample|AlwaysSample) means using 100%/0% probability and 0 lower bound rate

### Before
```
AdaptiveNeverSampleNoLateBinding/children=false/cpus-32         	  262648	      3886 ns/op
AdaptiveNeverSampleNoLateBinding/children=true/cpus-32          	 2388068	       485 ns/op
AdaptiveAlwaysSampleNoLateBinding/children=false/cpus-32        	  486133	      4813 ns/op
AdaptiveAlwaysSampleNoLateBinding/children=true/cpus-32         	 1451222	       802 ns/op
AdaptiveNeverSampleWithLateBinding/children=false/cpus-32       	  453198	      3879 ns/op
AdaptiveNeverSampleWithLateBinding/children=true/cpus-32        	 2221748	       489 ns/op
AdaptiveAlwaysSampleWithLateBinding/children=false/cpus-32      	  343279	      4946 ns/op
AdaptiveAlwaysSampleWithLateBinding/children=true/cpus-32       	 1348515	       827 ns/op
```

### After
```
NeverSample/children=false/cpus-32         	 3243393	       408 ns/op
NeverSample/children=true/cpus-32          	 2363767	       500 ns/op
AlwaysSample/children=false/cpus-32        	 1700967	       748 ns/op
AlwaysSample/children=true/cpus-32         	 1424816	       811 ns/op
AdaptiveNeverSampleNoLateBinding/children=false/cpus-32         	  561810	      2877 ns/op
AdaptiveNeverSampleNoLateBinding/children=true/cpus-32          	 2294772	       507 ns/op
AdaptiveAlwaysSampleNoLateBinding/children=false/cpus-32        	  404061	      3842 ns/op
AdaptiveAlwaysSampleNoLateBinding/children=true/cpus-32         	 1350417	       832 ns/op
AdaptiveNeverSampleWithLateBinding/children=false/cpus-32       	  455553	      4285 ns/op
AdaptiveNeverSampleWithLateBinding/children=true/cpus-32        	  260630	      3870 ns/op
AdaptiveAlwaysSampleWithLateBinding/children=false/cpus-32      	  265951	      4425 ns/op
AdaptiveAlwaysSampleWithLateBinding/children=true/cpus-32       	  255801	      4293 ns/op
```